+++ CHANGELOG.md
@@ -8,9 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 List all changes after the last release here (newer on top). Each change on a separate bullet point line
 
-### Fixed
+### Changed
 
-- Core: fix service components get cost
+- Core: only consider lines from the same supplier as the behavior component
 
 ### Fixed